# test: Sort tested dimensions dict by keys on A/B test failures