## Changes
- Sort the tested dimension dict by keys on A/B test failures using [`sort_keys=True`](https://docs.python.org/3/library/json.html#json.dump).


## Reason

When A/B test fails, it prints the tested dimensions, but it was not sorted and was not easy for human to read.

For example,
```
[Firecracker A/B-Test Runner] A/B-testing shows a change of 9.41MB/s, or 1.02%, (from 920.02MB/s to 929.44MB/s) for metric bw_write with p=0.00019998000199980003. This means that observing a change of this magnitude or worse, assuming that performance characteristics did not change across the tested commits, has a probability of 0.02%. Tested Dimensions:
{
 "io_engine": "Async",
 "guest_kernel": "linux-4.14",
 "host_kernel": "linux-5.10",
 "rootfs": "ubuntu-22.04.squashfs",
 "fio_mode": "randwrite",
 "cpu_model": "Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz",
 "guest_memory": "1024.0MB",
 "instance": "c5n.metal",
 "performance_test": "test_block_performance",
 "vcpus": "2",
 "fio_block_size": "4096"
}
[Firecracker A/B-Test Runner] A/B-testing shows a change of 4.10MB/s, or 1.22%, (from 335.69MB/s to 339.79MB/s) for metric bw_write with p=0.00019998000199980003. This means that observing a change of this magnitude or worse, assuming that performance characteristics did not change across the tested commits, has a probability of 0.02%. Tested Dimensions:
{
 "host_kernel": "linux-5.10",
 "guest_kernel": "linux-5.10",
 "rootfs": "ubuntu-22.04.squashfs",
 "fio_mode": "randwrite",
 "cpu_model": "Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz",
 "guest_memory": "1024.0MB",
 "instance": "c5n.metal",
 "performance_test": "test_block_performance",
 "io_engine": "Sync",
 "vcpus": "1",
 "fio_block_size": "4096"
}
```

This change make it look like this:
```
[Firecracker A/B-Test Runner] A/B-testing shows a change of 9.41MB/s, or 1.02%, (from 920.02MB/s to 929.44MB/s) for metric bw_write with p=0.00019998000199980003. This means that observing a change of this magnitude or worse, assuming that performance characteristics did not change across the tested commits, has a probability of 0.02%. Tested Dimensions:
{
  "cpu_model": "Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz",
  "fio_block_size": "4096",
  "fio_mode": "randwrite",
  "guest_kernel": "linux-4.14",
  "guest_memory": "1024.0MB",
  "host_kernel": "linux-5.10",
  "instance": "c5n.metal",
  "io_engine": "Async",
  "performance_test": "test_block_performance",
  "rootfs": "ubuntu-22.04.squashfs",
  "vcpus": "2"
}
[Firecracker A/B-Test Runner] A/B-testing shows a change of 4.10MB/s, or 1.22%, (from 335.69MB/s to 339.79MB/s) for metric bw_write with p=0.00019998000199980003. This means that observing a change of this magnitude or worse, assuming that performance characteristics did not change across the tested commits, has a probability of 0.02%. Tested Dimensions:
{
  "cpu_model": "Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz",
  "fio_block_size": "4096",
  "fio_mode": "randwrite",
  "guest_kernel": "linux-5.10",
  "guest_memory": "1024.0MB",
  "host_kernel": "linux-5.10",
  "instance": "c5n.metal",
  "io_engine": "Sync",
  "performance_test": "test_block_performance",
  "rootfs": "ubuntu-22.04.squashfs",
  "vcpus": "1"
}
```

## License Acceptance

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license. For more information on following Developer
Certificate of Origin and signing off your commits, please check
[`CONTRIBUTING.md`][3].

## PR Checklist

- ~~[ ] If a specific issue led to this PR, this PR closes the issue.~~
- [x] The description of changes is clear and encompassing.
- ~~[ ] Any required documentation changes (code and docs) are included in this
  PR.~~
- ~~[ ] API changes follow the [Runbook for Firecracker API changes][2].~~
- ~~[ ] User-facing changes are mentioned in `CHANGELOG.md`.~~
- [x] All added/changed functionality is tested.
- ~~[ ] New `TODO`s link to an issue.~~
- [x] Commits meet
  [contribution quality standards](https://github.com/firecracker-microvm/firecracker/blob/main/CONTRIBUTING.md#contribution-quality-standards).

______________________________________________________________________

- [x] This functionality cannot be added in [`rust-vmm`][1].

[1]: https://github.com/rust-vmm
[2]: https://github.com/firecracker-microvm/firecracker/blob/main/docs/api-change-runbook.md
[3]: https://github.com/firecracker-microvm/firecracker/blob/main/CONTRIBUTING.md